# expression: implement vectorized evaluation for `builtinJSONLengthSig`