--- a/expression/builtin_json_vec_test.go
+++ b/expression/builtin_json_vec_test.go
@@ -28,11 +28,15 @@ var vecBuiltinJSONCases = map[string][]vecExprBenchCase{
 	ast.JSONArrayAppend:  {},
 	ast.JSONContainsPath: {},
 	ast.JSONExtract:      {},
-	ast.JSONLength:       {},
-	ast.JSONType:         {{retEvalType: types.ETString, childrenTypes: []types.EvalType{types.ETJson}}},
-	ast.JSONArray:        {},
-	ast.JSONArrayInsert:  {},
-	ast.JSONContains:     {},
+	ast.JSONLength: {
+		{retEvalType: types.ETInt, childrenTypes: []types.EvalType{types.ETJson}},
+		{retEvalType: types.ETInt, childrenTypes: []types.EvalType{types.ETJson, types.ETString}, geners: []dataGenerator{nil, &constStrGener{"$.key"}}},
+		{retEvalType: types.ETInt, childrenTypes: []types.EvalType{types.ETJson, types.ETString}, geners: []dataGenerator{nil, &constStrGener{"$.abc"}}},
+	},
+	ast.JSONType:        {{retEvalType: types.ETString, childrenTypes: []types.EvalType{types.ETJson}}},
+	ast.JSONArray:       {},
+	ast.JSONArrayInsert: {},
+	ast.JSONContains:    {},
 	ast.JSONObject: {
 		{
 			retEvalType: types.ETJson,