<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
implement vectorized evaluation for `builtinJSONLengthSig` at https://github.com/pingcap/tidb/issues/12104

### What is changed and how it works?
the performance is related to the test case. It will get faster from 1% to 6 times. 
```
BenchmarkVectorizedBuiltinJSONFunc/builtinJSONLengthSig-VecBuiltinFunc-12                 300879              3545 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinJSONFunc/builtinJSONLengthSig-NonVecBuiltinFunc-12               67075             17767 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinJSONFunc/builtinJSONLengthSig-VecBuiltinFunc#01-12                2026            525188 ns/op          301371 B/op       2410 allocs/op
BenchmarkVectorizedBuiltinJSONFunc/builtinJSONLengthSig-NonVecBuiltinFunc#01-12             2137            554590 ns/op          301741 B/op       2410 allocs/op
BenchmarkVectorizedBuiltinJSONFunc/builtinJSONLengthSig-VecBuiltinFunc#02-12                2059            519725 ns/op          302338 B/op       2422 allocs/op
BenchmarkVectorizedBuiltinJSONFunc/builtinJSONLengthSig-NonVecBuiltinFunc#02-12             2120            540376 ns/op          302575 B/op       2422 allocs/op
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test